[refactor] router & controller 更名並修改cypherQuery
