+++ routers/fridgeRouters.js
@@ -1,6 +1,6 @@
 import express from 'express';
 
-import * as fridge from '../controllers/fridge.js';
+import * as fridge from '../controllers/fridgeControllers.js';
 import { validateJWT, upload } from '../middleware/userMiddleware.js';
 import catchAsync from '../utils/catchAsync.js';
