+++ utils/socket.js
@@ -2,6 +2,7 @@ import { Server } from 'socket.io';
 import http from 'http';
 
 let onlineUsers = [];
+// eslint-disable-next-line import/no-mutable-exports
 let io = null;
 
 const addNewUser = (userId, groupId, socketId) =>
@@ -22,6 +23,7 @@ export const initializeSocket = (app) => {
   });
 
   io.on('connection', (socket) => {
+    // eslint-disable-next-line no-console
     console.log('user connected');
     socket.on('newUser', (userInfo) => {
       addNewUser(userInfo.userId, userInfo.groupId, socket.id);
